Time: 100 ms (22.7%), Space: 27.3 MB (31.24%) - LeetHub

+++ 1615-range-sum-of-sorted-subarray-sums/1615-range-sum-of-sorted-subarray-sums.cpp
@@ -16,6 +16,6 @@ public:
         for(int i =left-1;i<right;i++){
               res=(res+vec[i])%mod;
         }
-        return static_cast<int>(res);
+        return res;
     }
 };
